tld.tests.test_commands: remove
These test is just testing the wrapper created by setuptools, and it
seems to be rather complicated to add it to PATH inside a tox
environment.

As the wrapper points to tld.utils:update_tld_names, which we already
test in tld/tests/test_core.py, let's drop this file. We hopefully can
rely on setuptools to do its job :-)